+++ icandigit.lua
@@ -1,3 +1,5 @@
+local du = require "darkutils"
+
 -- Nether Ice Bridge Excavator
 -- v1.0
 
@@ -8,45 +10,13 @@
 local fuel = "betterend:charcoal_block"
 local foundation = "minecraft:netherrack"
 
--- will search for and select a block if found,
--- otherwise prompt user input
-local function grabSome(block)
-  -- short circuit if item is already selected
-  if turtle.getItemCount() > 0 then
-    if turtle.getItemDetail().name == block then
-      return true
-    end
-  end
-  
-  -- No luck. Gotta search
-  for i=1, 16, 1 do
-    turtle.select(i)
-    if turtle.getItemCount() > 0 then
-      if turtle.getItemDetail().name == block then
-        return true
-      end
-    end
-  end
-  print("We're out of "..block.."!")
-  print("Please insert "..block.." then press [Enter] to continue.")
-  boob = io.read()
-  grabSome(block)
-end
-
-local function fuelCheck()
-  if turtle.getFuelLevel() < 80 then
-    grabSome(fuel)
-    turtle.refuel(1)
-  end
-end
-
 -- optionally digs up and places foundation underneath
 -- if block below is not foundation
 local function placeFoundation()
   local isablock, theblock = turtle.inspectDown()
     if not((theblock.name == foundation)) then
       turtle.digDown()
-      grabSome(foundation)
+      du.grabSome(foundation)
       turtle.placeDown()
   end
 end
@@ -120,11 +90,12 @@ local function dig()
 end
 
 local function main()
+  du.clearScreen()
   print("Hello and welcome to the nether bridge excavator v1")
   print("How far should I dig?")
   local distance = tonumber(io.read())
   for i=1, distance, 1 do
-    fuelCheck()
+    du.fuelCheck()
     dig()
   end
 end
